Rpc calls now have the option to be ordered or unordered

+++ GameEngine/include/GameEngine/Game_Manager.h
@@ -1,7 +1,7 @@
 #pragma once
 #include "Application.h"
 
-typedef uint32 Player_ID;
+typedef long Player_ID;
 
 class Game_Manager {
   private:
